refactor: move firebase logic to dedicated module
