+++ firebase/firebaseConfig.example.ts
