+++ hooks/useSignin.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { signInWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../firebaseConfig";
+import { auth } from "../firebase/firebaseConfig";
 
 export function useSignin() {
   const [loading, setLoading] = useState(false);
