+++ types/IUserData.ts
@@ -1,0 +1,4 @@
+export interface IUserData {
+  firstname: string;
+  surname: string;
+}
